pyproject.toml
[project]
name = "auth0-m2m-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
